requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
tensorflow==2.13.0
pillow==10.3.0
numpy==1.24.4
python-socketio==5.11.2
python-engineio==4.8.2
eventlet==0.33.3
pytz==2024.1
